Synchronize with the upstream
Note that it uses CRLF and WINDOWS-1251. So if we convert it we
lose a capability to track the upsream changes. So it's better
not to do utf-8/lf conversion unless the upstream chooses to do so.

Most non-Windows tools these days including git are perfectly
capable of working with this Windows native representation.